Config cleanup
