+++ src/main/kotlin/gui/dashboard/tiles/spectrum/SpectrumGuiConfig.kt
@@ -1,0 +1,7 @@
+package gui.dashboard.tiles.spectrum
+
+data class SpectrumGuiConfig(
+    val scale: Float,
+    val lowestFrequency: Float,
+    val highestFrequency: Float,
+)
